ci: add logging
ci: add logging

refactor: cloud-init to use data source

refactor: rename tftpl files
