+++ apps/web/pages/api/auth/[...auth0].ts
@@ -65,7 +65,9 @@ export default auth0Instance().handleAuth({
     }
   },
   login: async (req: NextApiRequest, res: NextApiResponse) => {
-    await auth0Instance().handleLogin(req, res, {
+    console.log('...auth0 login')
+
+    return auth0Instance().handleLogin(req, res, {
       returnTo: new URL('/apps', getEnv().auth0.baseUrl).toString(),
     })
   },
