--- libs/backend/infra/adapter/graphql/src/proxy.middleware.ts
@@ -1,16 +1,0 @@
-import { getEnv } from '@codelab/shared/config'
-import type { NextApiHandler } from 'next'
-import httpProxyMiddleware from 'next-http-proxy-middleware'
-import { authMiddleware } from './auth.middleware'
-import { corsMiddleware } from './cors.middleware'
-
-/**
- * Proxy requests to api
- */
-export const proxyMiddleware: NextApiHandler = async (req, res) => {
-  await corsMiddleware(req, res)
-  await authMiddleware(req, res)
-  await httpProxyMiddleware(req, res, {
-    target: getEnv().endpoint.apiHost,
-  })
-}
