+++ terraform/environments/prod/module-codelab-web.tf
@@ -4,6 +4,8 @@ module "codelab_web" {
   digitalocean_access_token = var.DIGITALOCEAN_ACCESS_TOKEN
   docker_tag_version        = var.DOCKER_TAG_VERSION
 
+  loki_url = var.LOKI_URL
+
   next_public_web_host = var.NEXT_PUBLIC_WEB_HOST
 
   next_public_api_hostname = module.codelab_api.codelab_api_hostname
