+++ terraform/environments/prod/vars-loki.tf
@@ -1,0 +1,3 @@
+variable "LOKI_URL" {
+  type = string
+}
