+++ terraform/environments/test/main.tf
@@ -4,7 +4,7 @@ terraform {
   required_providers {
     auth0 = {
       source  = "auth0/auth0"
-      version = "0.50.2"
+      version = "1.2.0"
     }
   }
