+++ terraform/modules/auth0/action_upsert_user.tf
@@ -22,6 +22,8 @@ const axios = require('axios');
   * @param {PostLoginAPI} api - Interface whose methods can be used to change the behavior of the login.
  */
 exports.onExecutePostLogin = async (event, api) => {
+  console.log(event)
+
   /**
    * Get Access Token
    */
@@ -38,6 +40,8 @@ exports.onExecutePostLogin = async (event, api) => {
   }
 
   const { data } = await axios.request(options)
+
+  console.log(data)
 
   /**
    * Initialize client
