+++ terraform/modules/auth0/client_machine.tf
@@ -11,7 +11,7 @@ resource "auth0_client" "machine_client" {
 resource "auth0_client_grant" "machine_client_grant" {
   client_id = auth0_client.machine_client.id
   audience  = local.auth0_audience
-  scope     = ["update:users"]
+  scopes    = ["update:users"]
 }
 
 data "auth0_client" "machine_client" {
