+++ terraform/modules/auth0/providers.tf
@@ -9,7 +9,7 @@ terraform {
   required_providers {
     auth0 = {
       source  = "auth0/auth0"
-      version = "0.50.2"
+      version = "1.2.0"
     }
   }
 }
