+++ terraform/modules/codelab-api/certificate.tf
@@ -1,0 +1,5 @@
+# resource "digitalocean_certificate" "api" {
+#   name = "codelab-api-certificate"
+#   type = "custom"
+#   # domains = ["example.com"]
+# }
