+++ terraform/modules/codelab-api/tftpl/daemon.tftpl.json
@@ -1,0 +1,7 @@
+{
+  "log-driver": "loki",
+  "log-opts": {
+    "loki-url": "${loki_url}",
+    "loki-batch-size": "400"
+  }
+}
