+++ terraform/modules/codelab-api/tftpl/docker-compose.tftpl.yaml
@@ -1,0 +1,15 @@
+version: '3.8'
+services:
+  api:
+    image: registry.digitalocean.com/codelabapp/api:${docker_tag_version}
+    container_name: codelab-api
+    restart: unless-stopped
+    ports:
+      - '${next_public_api_port}:4000'
+    environment:
+      - AUTH0_M2M_CLIENT_ID=${auth0_m2m_client_id}
+      - AUTH0_M2M_CLIENT_SECRET=${auth0_m2m_client_secret}
+      - AUTH0_DOMAIN=${auth0_domain}
+      - NEO4J_URI=${neo4j_uri}
+      - NEO4J_USER=${neo4j_user}
+      - NEO4J_PASSWORD=${neo4j_password}
