+++ terraform/modules/codelab-api/vars-loki-url.symlink.tf
@@ -1,0 +1,1 @@
+../loki/shared/vars-loki-url.tf
