--- terraform/modules/codelab-landing/landing-droplet.yaml
@@ -1,60 +1,0 @@
-#cloud-config
-
-# lifecycle in /etc/cloud/cloud.cfg
-
-# cloud-init config in /var/lib/cloud/instance/user-data.txt
-
-# /var/lib/cloud/instances/*/sem tells us which has been ran
-
-# manage_resolv_conf: true
-
-# package_update: true
-# package_upgrade: true
-# package_reboot_if_required: true
-
-write_files:
-  - path: /root/.config/doctl/config.yaml
-    content: |
-      access-token: ${digitalocean_access_token}
-  - path: /root/docker/docker-compose.yml
-    # owner: docker:docker
-    permissions: '0644'
-    content: |
-      version: '3.8'
-      services:
-        landing:
-          image: registry.digitalocean.com/codelabapp/landing:${docker_tag_version}
-          restart: unless-stopped
-          ports:
-            - "80:3000"
-    # triggered by config_scripts_user
-  - path: /var/lib/cloud/scripts/per-once/start-docker-image.sh
-    permissions: '0755'
-    # Indent issue for multiline
-    #
-    # https://discuss.hashicorp.com/t/wrong-indent-with-multiline-content-to-cloud-init-write-files-content-directives/35011/3
-    # https://github.com/hashicorp/terraform/issues/10378#issuecomment-263151581
-    #
-    # Also cannot have spaces up front https://stackoverflow.com/questions/23818209/user-data-scripts-fails-without-giving-reason
-    content: |
-      #!/bin/bash
-      set -x
-
-      snap install doctl
-
-      snap connect doctl:dot-docker
-
-      doctl auth init --interactive false --access-token ${digitalocean_access_token}
-
-      doctl registry login
-
-      cd /root/docker && docker compose pull && docker compose up -d --force-recreate
-
-# Still need to use this to call the user scripts
-runcmd:
-  - [
-      cloud-init-per,
-      once,
-      start-docker-image,
-      /var/lib/cloud/scripts/per-once/start-docker-image.sh,
-    ]
