+++ terraform/modules/codelab-neo4j/codelab-neo4j-droplet.tf
@@ -14,9 +14,7 @@ resource "digitalocean_droplet" "neo4j" {
   ssh_keys = ["31:0e:90:12:06:a2:9f:8b:07:0e:a8:49:cc:d8:1f:71"]
 
   # Run once only
-  user_data = templatefile("${path.module}/codelab-neo4j-droplet.yaml", {
-    digitalocean_access_token = var.digitalocean_access_token,
-  })
+  user_data = data.cloudinit_config.neo4j.rendered
 
   lifecycle {
     # ignore_changes = [user_data]
@@ -30,5 +28,7 @@ resource "digitalocean_droplet" "neo4j" {
 }
 
 output "neo4j_uri" {
-  value = "neo4j+s://${digitalocean_droplet.neo4j.ipv4_address_private}:7687"
+  # https://stackoverflow.com/questions/62357682/routing-issue-in-neo4j-4-0-with-multiple-databases
+  value = "bolt://${digitalocean_droplet.neo4j.ipv4_address_private}:7687"
+  # value = "bolt+s://${digitalocean_droplet.neo4j.ipv4_address_private}:7687"
 }
