--- terraform/modules/codelab-neo4j/codelab-neo4j-droplet.yaml
@@ -1,86 +1,0 @@
-#cloud-config
-
-# lifecycle in /etc/cloud/cloud.cfg
-
-# cloud-init config in /var/lib/cloud/instance/user-data.txt
-
-# /var/lib/cloud/instances/*/sem tells us which has been ran
-
-# manage_resolv_conf: true
-
-# package_update: true
-# package_upgrade: true
-# package_reboot_if_required: true
-
-write_files:
-  - path: /root/.config/doctl/config.yaml
-    content: |
-      access-token: ${digitalocean_access_token}
-  - path: /root/docker/docker-compose.yml
-    # owner: docker:docker
-    permissions: '0644'
-    content: |
-      version: '3.8'
-      services:
-        neo4j:
-          container_name: neo4j
-          image: neo4j:5.11.0
-          volumes:
-            - codelab-neo4j:/data
-            - codelab-neo4j:/plugins
-            - codelab-neo4j:/import
-            - codelab-neo4j:/logs
-          ports:
-            - '7474:7474'
-            - '7687:7687'
-          environment:
-            - NEO4J_ACCEPT_LICENSE_AGREEMENT=yes
-            - NEO4J_apoc_export_file_enabled=true
-            - NEO4J_apoc_import_file_enabled=true
-            - NEO4J_apoc_import_file_use__neo4j__config=true
-            # https://neo4j.com/docs/operations-manual/current/configuration/configuration-settings/#config_server.default_listen_address
-            - NEO4J_server_default__listen__address=0.0.0.0
-            - NEO4J_dbms.security.procedures.unrestricted=apoc.*
-            - NEO4J_PLUGINS=["apoc"]
-            - NEO4J_AUTH=neo4j/password
-          # networks: ["codelab-neo4j"]
-          restart: unless-stopped
-      volumes:
-        codelab-neo4j:
-          name: codelab-neo4j
-          driver: local
-          driver_opts:
-            type: none
-            o: bind
-            device: /mnt/neo4j_volumes
-    # triggered by config_scripts_user
-  - path: /var/lib/cloud/scripts/per-once/start-docker-image.sh
-    permissions: '0755'
-    # Indent issue for multiline
-    #
-    # https://discuss.hashicorp.com/t/wrong-indent-with-multiline-content-to-cloud-init-write-files-content-directives/35011/3
-    # https://github.com/hashicorp/terraform/issues/10378#issuecomment-263151581
-    #
-    # Also cannot have spaces up front https://stackoverflow.com/questions/23818209/user-data-scripts-fails-without-giving-reason
-    content: |
-      #!/bin/bash
-      set -x
-
-      snap install doctl
-
-      snap connect doctl:dot-docker
-
-      doctl auth init --interactive false --access-token ${digitalocean_access_token}
-
-      doctl registry login
-
-      cd /root/docker && docker compose pull && docker compose up -d --force-recreate
-
-# Still need to use this to call the user scripts
-runcmd:
-  - [
-      cloud-init-per,
-      once,
-      start-docker-image,
-      /var/lib/cloud/scripts/per-once/start-docker-image.sh,
-    ]
