+++ terraform/modules/codelab-neo4j/tftpl/docker-compose.tftpl.yaml
@@ -1,0 +1,33 @@
+version: '3.8'
+services:
+  neo4j:
+    image: neo4j:5.11.0
+    container_name: neo4j
+    volumes:
+      - codelab-neo4j:/data
+      # - codelab-neo4j:/plugins
+      # - codelab-neo4j:/import
+      # - codelab-neo4j:/logs
+    ports:
+      - '7474:7474'
+      - '7687:7687'
+    environment:
+      - NEO4J_ACCEPT_LICENSE_AGREEMENT=yes
+      - NEO4J_apoc_export_file_enabled=true
+      - NEO4J_apoc_import_file_enabled=true
+      - NEO4J_apoc_import_file_use__neo4j__config=true
+      # https://neo4j.com/docs/operations-manual/current/configuration/configuration-settings/#config_server.default_listen_address
+      - NEO4J_server_default__listen__address=0.0.0.0
+      - NEO4J_dbms.security.procedures.unrestricted=apoc.*
+      - NEO4J_PLUGINS=["apoc"]
+      - NEO4J_AUTH=neo4j/password
+    # networks: ["codelab-neo4j"]
+    restart: unless-stopped
+volumes:
+  codelab-neo4j:
+    name: codelab-neo4j
+    driver: local
+    driver_opts:
+      type: none
+      o: bind
+      device: /mnt/neo4j_volumes
