--- terraform/modules/codelab-web/codelab-web-droplet.yaml
@@ -1,63 +1,0 @@
-#cloud-config
-
-# lifecycle in /etc/cloud/cloud.cfg
-
-# cloud-init config in /var/lib/cloud/instance/user-data.txt
-
-# /var/lib/cloud/instances/*/sem tells us which has been ran
-
-# manage_resolv_conf: true
-
-# package_update: true
-# package_upgrade: true
-# package_reboot_if_required: true
-
-write_files:
-  - path: /root/.config/doctl/config.yaml
-    content: |
-      access-token: ${digitalocean_access_token}
-  - path: /root/docker/docker-compose.yml
-    # owner: docker:docker
-    permissions: '0644'
-    #
-    content: |
-      version: '3.8'
-      services:
-        web:
-          image: registry.digitalocean.com/codelabapp/web:${docker_tag_version}
-          restart: unless-stopped
-          ports:
-            - "80:3000"
-          environment:
-            - NEXT_PUBLIC_WEB_HOST=${next_public_web_host}
-            - NEXT_PUBLIC_API_PORT=${next_public_api_port}
-            - NEXT_PUBLIC_API_HOSTNAME=${next_public_api_hostname}
-            - AUTH0_SECRET=${auth0_secret}
-            - AUTH0_DOMAIN=${auth0_domain}
-            - AUTH0_CLIENT_ID=${auth0_client_id}
-            - AUTH0_CLIENT_SECRET=${auth0_client_secret}
-    # triggered by config_scripts_user
-  - path: /var/lib/cloud/scripts/per-once/start-docker-image.sh
-    permissions: '0755'
-    content: |
-      #!/bin/bash
-      set -x
-
-      snap install doctl
-
-      snap connect doctl:dot-docker
-
-      doctl auth init --interactive false --access-token ${digitalocean_access_token}
-
-      doctl registry login
-
-      cd /root/docker && docker compose pull && docker compose up -d --force-recreate
-
-# Still need to use this to call the user scripts
-runcmd:
-  - [
-      cloud-init-per,
-      once,
-      start-docker-image,
-      /var/lib/cloud/scripts/per-once/start-docker-image.sh,
-    ]
