+++ terraform/modules/loki/shared/vars-loki-url.tf
@@ -1,0 +1,3 @@
+variable "loki_url" {
+  type = string
+}
